requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.2
chromadb==0.4.7
click==8.1.7
colorclass==2.2.2
coloredlogs==15.0.1
compressed-rtf==1.0.6
cryptography==42.0.5
dataclasses-json==0.5.14
diskcache==5.6.3
easygui==0.98.3
ebcdic==1.1.1
emoji==2.10.1
exceptiongroup==1.2.0
extract-msg==0.45.0
fastapi==0.99.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
gpt4all==1.0.8
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.21.3
humanfriendly==10.0
idna==3.6
IMAPClient==2.3.1
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.3.2
langchain==0.0.274
langsmith==0.0.92
lark==1.1.8
llama_cpp_python==0.1.81
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.21.0
monotonic==1.6
mpmath==1.3.0
msoffcrypto-tool==5.3.1
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
olefile==0.46
oletools==0.60.1
onnxruntime==1.17.1
overrides==7.7.0
packaging==23.2
pandoc==2.3
pcodedmp==1.2.6
pdfminer.six==20221105
pillow==10.2.0
plumbum==1.8.2
ply==3.11
posthog==3.5.0
protobuf==4.25.3
pulsar-client==3.4.0
pycparser==2.21
pydantic==1.10.14
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pypandoc==1.11
pyparsing==2.4.7
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.9
PyYAML==6.0.1
red-black-tree-mod==1.20
regex==2023.12.25
requests==2.31.0
RTFDE==0.1.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.1
transformers==4.38.2
triton==2.2.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzlocal==5.2
unstructured==0.10.8
urllib3==2.0.4
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
